+++ src/jhson.js
@@ -18,6 +18,7 @@
         _parameter_Document = null,
         _parameter_Window = null,
         _parameter_JSON = null,
+        _parameter_Math = null,
 
         // Variables: Configuration
         _configuration = {},
@@ -28,7 +29,8 @@
         // Variables: Strings
         _string = {
             empty: "",
-            space: " "
+            space: " ",
+            newLine: "\n"
         },
 
         // Variables: JSON
@@ -231,7 +233,7 @@
                             element.innerHTML = _string.empty;
                         }
 
-                        writeNode( element, convertedJsonObject.result[ key ], templateDataKeys, properties.templateData );
+                        writeNode( element, convertedJsonObject.result[ key ], templateDataKeys, properties );
                     }
                 }
             }
@@ -240,8 +242,9 @@
         return _this;
     }
 
-    function writeNode( element, jsonObject, templateDataKeys, templateData ) {
-        var templateDataKeysLength = templateDataKeys.length;
+    function writeNode( element, jsonObject, templateDataKeys, properties ) {
+        var templateDataKeysLength = templateDataKeys.length,
+            cssStyles = [];
 
         for ( var jsonKey in jsonObject ) {
             if ( startsWithAnyCase( jsonKey, _json.attribute ) ) {
@@ -253,7 +256,11 @@
             } else if ( startsWithAnyCase( jsonKey, _json.cssStyle ) ) {
                 var cssStyleName = jsonKey.replace( _json.cssStyle, _string.empty );
 
-                element.style[ cssStyleName ] = jsonObject[jsonKey];
+                if ( !properties.writeCssToHead ) {
+                    element.style[ cssStyleName ] = jsonObject[ jsonKey ];
+                } else {
+                    cssStyles.push( cssStyleName + ":" + jsonObject[ jsonKey ] + ";" );
+                }
 
             } else if ( jsonKey === _json.text ) {
                 element.innerHTML = jsonObject[ jsonKey ];
@@ -262,8 +269,8 @@
                     for ( var templateDataKeyIndex = 0; templateDataKeyIndex <  templateDataKeysLength; templateDataKeyIndex++ ) {
                         var templateDataKey = templateDataKeys[ templateDataKeyIndex ];
 
-                        if ( templateData.hasOwnProperty( templateDataKey ) ) {
-                            element.innerHTML = replaceAll( element.innerHTML, templateDataKey, templateData[ templateDataKey ] );
+                        if ( properties.templateData.hasOwnProperty( templateDataKey ) ) {
+                            element.innerHTML = replaceAll( element.innerHTML, templateDataKey, properties.templateData[ templateDataKey ] );
                         }
                     }
                 }
@@ -278,12 +285,33 @@
                         if ( childJson.hasOwnProperty( childJsonKey ) ) {
                             var childElement = createElement( element, childJsonKey.toLowerCase() );
 
-                            writeNode( childElement, childJson[ childJsonKey ], templateDataKeys, templateData );
+                            writeNode( childElement, childJson[ childJsonKey ], templateDataKeys, properties );
                         }
                     }
                 }
             }
         }
+
+        if ( cssStyles.length > 0 ) {
+            writeCssStyleTag( element, cssStyles );
+        } 
+    }
+
+    function writeCssStyleTag( element, cssStyles ) {
+        var head = _parameter_Document.getElementsByTagName( "head" )[ 0 ];
+
+        if ( !isDefinedString( element.id ) ) {
+            element.id = newGuid();
+        }
+
+        var cssLines = [];
+        cssLines.push( "#" + element.id + " {" );
+        cssLines = cssLines.concat( cssStyles );
+        cssLines.push( "}" );
+
+        var style = createElement( head, "style" );
+        style.type = "text/css";
+        style.appendChild( _parameter_Document.createTextNode( cssLines.join( _string.newLine ) ) );
     }
 
 
@@ -349,6 +377,21 @@
      * String Handling
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
      */
+
+    function newGuid() {
+        var result = [];
+
+        for ( var charIndex = 0; charIndex < 32; charIndex++ ) {
+            if ( charIndex === 8 || charIndex === 12 || charIndex === 16 || charIndex === 20 ) {
+                result.push( _string.dash );
+            }
+
+            var character = _parameter_Math.floor( _parameter_Math.random() * 16 ).toString( 16 );
+            result.push( character );
+        }
+
+        return result.join( _string.empty );
+    }
 
     function startsWithAnyCase( data, start ) {
         return data.substring( 0, start.length ).toLowerCase() === start.toLowerCase();
@@ -594,7 +637,8 @@
                 json: _string.empty,
                 templateData: {},
                 removeAttributes: true,
-                clearHTML: true
+                clearHTML: true,
+                writeCssToHead: false
             };
 
             /**
@@ -661,6 +705,23 @@
              */
             scope.clearHTML = function( flag ) {
                 __properties.clearHTML = getDefaultBoolean( flag, __properties.clearHTML );
+
+                return this;
+            };
+
+            /**
+             * writeCssToHead().
+             * 
+             * States if the CSS style properties should be written to a "style" tag in the HTML documents HEAD DOM element.
+             * 
+             * @public
+             * 
+             * @param       {boolean}    flag                               The boolean flag that states the condition (defaults to false).
+             * 
+             * @returns     {Object}                                        The HTML properties object.
+             */
+            scope.writeCssToHead = function( flag ) {
+                __properties.writeCssToHead = getDefaultBoolean( flag, __properties.writeCssToHead );
 
                 return this;
             };
@@ -766,10 +827,11 @@
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
      */
 
-    ( function ( documentObject, windowObject, jsonObject ) {
+    ( function ( documentObject, windowObject, jsonObject, mathObject ) {
         _parameter_Document = documentObject;
         _parameter_Window = windowObject;
         _parameter_JSON = jsonObject;
+        _parameter_Math = mathObject;
 
         buildDefaultConfiguration();
 
@@ -777,5 +839,5 @@
             _parameter_Window.$jhson = this;
         }
 
-    } ) ( document, window, JSON );
+    } ) ( document, window, JSON, Math );
 } )();