+++ docs/CHANGE_LOG.md
@@ -1,7 +1,9 @@
 # JHson.js - Change Log:
 
 ## Version 0.4.0:
-- 
+
+#### **Public Functions:**
+- Added "writeCssToHead()" support for "html()", which will force all CSS style properties to be added as "style" tags to the head of the HTML document.
 
 <br>